+++ src/rinoh/frontend/sphinx/transforms.py
@@ -7,6 +7,8 @@ from sphinx.domains.citation import CitationDomain
 from sphinx.transforms.post_transforms import SphinxPostTransform
 from sphinx.util.nodes import NodeMatcher
 
+from .util import fully_qualified_id
+
 
 class RinohCitationReferenceTransform(SphinxPostTransform):
     default_priority = 5  # before ReferencesResolver
@@ -16,12 +18,14 @@ class RinohCitationReferenceTransform(SphinxPostTransform):
     def run(self, **kwargs) -> None:
         domain = cast(CitationDomain, self.env.get_domain('citation'))
         pending_xref = NodeMatcher(addnodes.pending_xref, refdomain='citation',
-                                    reftype='ref')
+                                   reftype='ref')
         for node in self.document.traverse(pending_xref):
-            doc, refid, _ = domain.citations.get(node['reftarget'], ('', '', 0))
-            if doc:
+            docname, reftarget, _ = domain.citations.get(node['reftarget'],
+                                                         ('', '', 0))
+            refid = fully_qualified_id(docname, reftarget)
+            if docname:
                 child = Text(node['reftarget'])
                 citation_ref = nodes.citation_reference('', '', child,
-                                                        docname=doc,
+                                                        docname=docname,
                                                         refid=refid)
                 node.replace_self(citation_ref)